--- public/index.html
@@ -1,9 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <title>React App</title>
-  </head>
-  <body>
-    <div id="root"></div>
-  </body>
-</html>
