+++ scholarshipper/package.json
@@ -22,12 +22,6 @@
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.5.0",
     "@testing-library/user-event": "^7.2.1",
-    "@types/jest": "^24.9.1",
-    "@types/node": "^12.12.43",
-    "@types/react": "^16.9.35",
-    "@types/react-dom": "^16.9.8",
-    "apollo-client": "^2.6.10",
-    "apollo-server-express": "^2.14.1",
     "axios": "^0.19.2",
     "bcryptjs": "^2.4.3",
     "bootstrap": "^4.5.0",
@@ -49,8 +43,6 @@
     "react-router-dom": "^5.2.0",
     "react-scripts": "3.4.1",
     "reflect-metadata": "^0.1.13",
-    "typeorm": "^0.2.25",
-    "typescript": "^3.9.3",
     "uuid": "^8.1.0"
   },
   "scripts": {
@@ -59,20 +51,12 @@
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject",
-    "dev": "nodemon server/server.ts",
     "package-mac": "electron-packager . --overwrite --platform=darwin --arch=x64 --icon=assets/icons/mac/icon.icns --prune=false --out=release-builds",
     "package-win": "electron-packager . electron-tutorial-app --overwrite --asar=true --platform=win32 --arch=ia32 --icon=assets/icons/win/icon.ico --prune=false --out=release-builds --version-string.CompanyName=CE --version-string.FileDescription=CE --version-string.ProductName=\"ScholarShipper\"",
     "package-linux": "electron-packager . electron-tutorial-app --overwrite --asar=true --platform=linux --arch=x64 --icon=assets/icons/png/icon.png --prune=false --out=release-builds"
   },
   "proxy": "http://localhost:3000/",
   "secure": false,
-  "eslintConfig": {
-    "extends": "react-app",
-    "@typescript-eslint/no-unused-vars": "off"
-  },
-  "eslintIgnore": [
-    "src/*"
-  ],
   "browserslist": {
     "production": [
       ">0.2%",
@@ -91,18 +75,12 @@
     "@babel/plugin-transform-modules-commonjs": "^7.10.1",
     "@babel/preset-env": "^7.0.0",
     "@babel/preset-react": "^7.0.0",
-    "@types/bcrypt": "^3.0.0",
-    "@types/express": "^4.17.6",
-    "@types/express-session": "^1.17.0",
-    "@types/graphql": "^14.5.0",
     "babel-eslint": "^10.0.0",
     "babel-loader": "^8.0.0",
     "dotenv": "^8.2.0",
     "electron-builder": "^22.7.0",
     "electron-packager": "^14.2.1",
     "electron-reload": "^1.5.0",
-    "nodemon": "^2.0.4",
-    "ts-node": "^8.10.2",
-    "typescript": "^3.9.3"
+    "nodemon": "^2.0.4"
   }
 }
