+++ scholarshipper/public/index.html
@@ -42,27 +42,6 @@
               ul.className = '';
             }
           };
-      </script>
-      <!-- <br></br>
-        <div class="card" style="width: 18rem;">
-          <div class="card-body">
-            <h5 class="card-title">Cohort</h5>
-            <p class="card-text"># goes here</p>
-            <a class="btn btn-primary btn-sm btn-info" href="#" role="button" onClick="studentPopUp()">Link</a>
-          </div>
-        </div>
-      <br></br>  -->
-      <script>
-      //   function studentPopUp(){
-      //       //.....
-      //       //assuming the authentication is valid, we want to show now the new window which will be our main window
-      //       const { ipcRenderer } = require('electron'); //require electron
-
-      //       //using ipcRenderer, let's call the function in index.js where the function name is `newWindow`, 
-      //       //the `main.html` is the new html file I want to show, use your own.
-      //       ipcRenderer.send('studentWindow','./Student.tsx'); 
-      //   }
-
-      // </script>
+        </script>
   </body>
 </html>
