--- scholarshipper/release-builds/scholarshippers-darwin-x64/version
@@ -1,1 +1,0 @@
-9.0.2
