--- scholarshipper/server/models/models.ts
@@ -1,14 +1,0 @@
-// const { Pool } = require('pg');
-
-// const POSTGRES = process.env.PG_URI;
-
-// const pool = new Pool({
-//   connectionString: POSTGRES
-// })
-
-// module.exports = {
-//   query: (text, params, callback) => {
-//     console.log('executed query', text);
-//     return pool.query(text, params, callback);
-//   }
-// };
