+++ scholarshipper/src/components/AddLogItem.jsx
@@ -4,8 +4,8 @@ import Form from 'react-bootstrap/Form';
 import Row from 'react-bootstrap/Row';
 import Col from 'react-bootstrap/Col';
 import Button from 'react-bootstrap/Button';
+// @ts-ignore
 const { ipcRenderer } = window.require('electron');
-
 
 const AddLogItem = ({ addItem }) => {
   const [notes, setNotes] = useState('');
