+++ scholarshipper/src/components/Card.jsx
@@ -1,24 +1,12 @@
 import React, { useEffect } from 'react';
 import { LinkContainer } from 'react-router-bootstrap';
-import Student from '../Student'
 const { ipcRenderer } = window.require('electron');
 
 const Card = (props) => {
-  // props.cohortData contains cohort_id and start_year
   const { cohortData } = props;
-  // const [students, setStudents] = useState([]);
-
-  // useEffect(() => {
-  //   // Initiate renderer process to ipcMain to query DB for cohort data.
-  //   ipcRenderer.send('getstudentCohort',['']);
-  //   // Catch the data sent back by ipcMain.
-  //   ipcRenderer.on('gotStudentCohorts', (event, cohortData) => {
-  //     setStudents([...students]);
-  //   })
-  // }, []);
 
   return (
-    <div className="card">
+    <div className="card" style={styles.card}>
       <div className="card-body">
         <h5 className="card-title">Cohort</h5>
         <LinkContainer to={`/Student/${cohortData.cohort_id}`}>
@@ -30,3 +18,11 @@ const Card = (props) => {
 }
 
 export default Card;
+
+const styles = {
+  card: {
+    padding: '2rem 4rem',
+    boxShadow: '0px 3px 5px rgba(0,0,0,0.1)',
+    marginBottom: '2rem'
+  }
+}
