+++ scholarshipper/src/components/Cohort.jsx
@@ -1,8 +1,8 @@
 import React, {useState,useEffect} from 'react';
-import './App.css';
-import logo from './assets/ScholarShipperIcon_.png';
-import Grid from './components/Grid.jsx';
 import { LinkContainer } from 'react-router-bootstrap';
+import logo from '../assets/ScholarShipperIcon_.png';
+import Grid from './Grid.jsx';
+import '../App.css';
 const { ipcRenderer } = window.require('electron'); 
 
 function Cohort() {
