+++ scholarshipper/src/components/Grid.jsx
@@ -1,7 +1,5 @@
 import React from 'react';
 import Container from 'react-bootstrap/Container';
-import Col from 'react-bootstrap/Col';
-import Row from 'react-bootstrap/Row';
 import Card from './Card.jsx';
 
 const Grid = (props) => {
