+++ scholarshipper/src/components/LogItem.jsx
