+++ scholarshipper/src/components/LogStudentInfo.jsx
