+++ scholarshipper/src/firstMain.js
@@ -2,8 +2,6 @@ const electron = require('electron')
 const path = require('path')
 const url = require('url')
 const isDev = require('electron-is-dev')
-
-
 require('electron-reload')
 const { app, BrowserWindow, Menu, ipcMain } = electron
