+++ scholarshipper/src/models/models.js
@@ -1,4 +1,3 @@
-// @ts-ignore
 const { Pool } = require('pg');
 require('dotenv').config();
